task: stub out mongoose connection and schema
