+++ data/db.ts
@@ -1,18 +1,30 @@
 import mongoose from "mongoose";
-export const MONGO_DB_URI = `mongodb+srv://${process.env.DB_USER}:${process.env.DB_PASSWORD}@${process.env.DB_SERVER}/?retryWrites=true&w=majority`;
-export const connect = (async () => await mongoose.connect(MONGO_DB_URI));
-export const userFavTokensSchema = new mongoose.Schema({
-    userId: 
-    {
-        type: String,
-        required: true,
-        unique: true
-    },
-    favTokenIds: {
-        type: [Number],
-        default:[]
+export const MONGODB_URI = `mongodb+srv://${process.env.DB_USER}:${process.env.DB_PASSWORD}@${process.env.DB_SERVER}/?retryWrites=true&w=majority`;
+let cached = { conn: null as any, promise: null as any };
+async function dbConnect() {
+    if (cached.conn) {
+      return cached.conn;
     }
-  });
+  
+    if (!cached.promise) {
+      const opts = {
+        bufferCommands: false,
+      };
+  
+      cached.promise = mongoose.connect(MONGODB_URI, opts);
+    }
+  
+    try {
+      cached.conn = await cached.promise;
+    } catch (e) {
+      cached.promise = null;
+      throw e;
+    }
+  
+    return cached.conn;
+  }
+  export default dbConnect;
+  
+  
 
-  // Create a model from the schema
-export const UserFavTokens = mongoose.model('UserFavTokens', userFavTokensSchema);
+
